requirements.txt
requests[socks]
netifaces
cysystemd
